Add a checker to encourage speeding up Error Prone checkers
VisitorState provides some methods to let you look up what javac knows
about something based on a String key. For variable strings, the best
you can do is call that method; but if you know ahead of time what
strings you'll be looking up, you can create a cache to avoid paying
the lookup cost.

One bug that may or may not be a feature: as written, I X-ray through
string constants, replacing

  final String CLASS = "java.lang.Class";
  ...
  state.getTypeFromString(CLASS)

with

  final String CLASS = "java.lang.Class";
  final Supplier<Type> JAVA_LANG_CLASS = VisitorState.memoize(
    s -> s.getTypeFromString("java.lang.Class"));
  ...
  JAVA_LANG_CLASS.get(state)

It would be nice to preserve the intent of the original CLASS
constant, but it's rather difficult to do if there may be different
use sites each using different constants (or no constant) for their
lookups. And in a way it's nice to be explicit about exactly what is
being looked up.

PiperOrigin-RevId: 330625691