Make ScopeOrQualifierAnnotationRetention postfix the Retention annotation
Retention is an added on top of @Qualifier. So @Qualifier is really the most important part we want to see first. Then comes the @Retention, that further describes what the annotation does.

Seeing Retention first just feels yucky to me: https://screenshot.googleplex.com/BcSHWHah6XjQtT8

This only makes it so the @Retention is added last, not “inserted after”:
- I think (with no data backing me up), that most of these annotations have only 1 annotation anyways.
- It would add what I think is unwarranted complexity by what could other be a 1 line fix.
- I wouldn’t know how to do it anyways.

PiperOrigin-RevId: 346513335