Update versions to HEAD-SNAPSHOT
following the example of e.g. Guava and Truth. Instead of pushing release and
SNAPSHOT versions to the main branch, we can commit them immediately before a
release and then push the tagged commit, but avoid cluttering up the main
branch or having to make the changes internally and wait for copybara.

PiperOrigin-RevId: 351872611