Provide a simple fix for `LockOnBoxedPrimitive`
that replaces locking on a boxed primitive field with locking on a
separate lock `Object`.

Also stop scanning the entire file for fields of type `Object` that
are initialized to boxed primitives: this seems to be rare in
practices, and is too slow to do if we want to enable this check at
build-time.

PiperOrigin-RevId: 353137559