Remove the `style` tag from MemberName and TestNameFormat.
While these are worth showing to CL authors (given they are style violations), they're likely not worth the noise of mailing fixes for.

Worse, confidence in the fixes is far from 100%, given they involve renaming things in a best-effort fashion. And even a small fraction of poor findings for a common pattern is a lot of bad hits.

PiperOrigin-RevId: 355811805