Disable UnnecessaryAnonymousClass by default
Method references aren't new anymore, and developers are generally aware
enough of them to use them as appropriate. Instead, we have found that this
check sometimes surprises developers, who were relying on reference equality
for multiple references to the same object, which is broken by extraction to a
method reference.

This also removes the hardcoded exemption for Android code. While it's true
that there are many more false positives on Android, it seems counterproductive
to have such an exemption for a check that is opt-in only: if an Android
codebase wants to opt into this check, we may as well let them.

PiperOrigin-RevId: 361229826