Rollforward of https://github.com/google/error-prone/commit/8c5883e0bc35fd7c049fdd2ef164ea8259a9f2a3: Add a `withNoParameters` method to Matchers.
Writing `withParameters` with no args makes it sound like you're saying it _does_ have parameters.

I guess due to fun versioning issues, we need to add the method and wait for a JavaBuilder release before using it.

Even more arcane, adding a `withParameters()` method in order to deprecate it changes the overload resolution for `withParameters()` calls which leads to a _different_ versioning issue.

PiperOrigin-RevId: 377259384