Convert `ReturnMissingNullable` to a `CompilationUnitTreeMatcher`.
This has various tradeoffs for the code as it stands today. But my motivation is to prepare to store information across `match*` calls so that we can identify `return` statements that return a `static final` field that is initialized to `null`.

Tests pass, including the new test (which I assume also passed before this CL).

PiperOrigin-RevId: 390008688