Don't add `@Nullable` to return types that are already `@PolyNull`.
While in the neighborhood, add some tests for the non-conservative mode.

Alternative implementation: It would be possible to enhance `NullnessAnnotations` to return separate values for `@Nullable` and `@PolyNull`. That would make a little more sense for `EqualsMissingNullable`, but who's going to write `boolean equals(@PolyNull Object o)`? :) This CL's current approach of treating the two identically probably will lead to the best overall results for callers, who mostly just want to know "Can this ever be null?"

PiperOrigin-RevId: 392024809