Introduce `VoidMissingNullable`.
I had initially figured that this refactoring wouldn't be very generally useful, since I was seeing `Void` mainly in code that uses javac visitor/scanner types. But I've since been reminded that we use `ListenableFuture<Void>` and similar types fairly often.

It's still hard to see much value in applying this refactoring depot-wide. But for people who are annotating their own projects by hand, it's nice for us to provide a tool to automate this very automatable part of the process.

PiperOrigin-RevId: 392977807