Automated rollback of commit 6804a8ea5a3f6d2a4ca72fbf78d13401589fc80f.
*** Reason for rollback ***

Seems to flag unintended code as string templates, see []

*** Original change description ***

Change `StrictFloggerFormatString` to ignore calls to log with a single parameter. `log(Object)` was already unaffected, but `log(String)` never should have been checked by it since the argument is not a format string. This change makes the behavior of the check match the behavior of the Java/Error-Prone version of the check (as demonstrated in the change to the Error-Prone check's test), meaning users migrating existing code from Java to Kotlin won't see new errors from this check.

It's still...

***

PiperOrigin-RevId: 393116658