[ #doc ] Document how to link to a method with a generic parameter in JavaDoc
There was no existing markdown document for InvalidLink, so I created one. The Oracle documentation doesn't make it clear that you need to omit the type parameter  (as far as I can see) and it was only by chance that I stumbled across the reason why my code was being flagged.

PiperOrigin-RevId: 395646736