Optionally allow a couple of methods in Java8ApiChecker
Disallowing these overrides is useful for code that expects to compile
with `-source 8 -target 8` and JDK 9 or newer, and then run on JDK 8.

Allowing the overrides is useful for code that wants to avoid using APIs
that are only exist in JDK 9 or newer, but that expects to be compiled
with a Java 8 bootclasspath before running on JDK 8.

PiperOrigin-RevId: 400757940