Add a single additional space in AnnotationPosition's fix.
Very occasionally we get a fix that formats to "... public@Nullable ..." and isn't fixed by the formatting we do when generating fixes. Curiously, it _is_ then fixed by formatting the file.

I've failed to write a repro test for this, but I reflumed and got no more of these issues.

PiperOrigin-RevId: 401748265