In ReturnMissingNullable, put type-use annotations in the right place.
This not only avoids...

- trying to annotate scoping constructs
- annotating the array element type instead of the array itself

...but also puts type-use annotations after modifiers (like AnnotationPosition) instead of before.

Along the way, I noticed a bug, which is that we always insert "@Nullable" when importing a type, even if that type is not named "Nullable." I fixed that and added a test.

Finally, I removed a TODO about @NullableDecl. We *could* use it if it's present, but I'd probably rather dedicate that effort to migrating off checker-compat-qual entirely.

(OK, and I also added a few more tests just to make sure that detection of an existing type-use @Nullable works.)

PiperOrigin-RevId: 403145907