Make ReturnMissingNullable recognize -XepOpt:Nullness:OnlyIfAnnotationAlreadyInScope=true.
By setting that flag, we can generate fixes only in code in which a nullness annotation is already imported (or declared in scope). This provides a conservative way to estimate "Is the nullness annotation available as a direct compile-time dependency?" -- and thus "Will this change actually compile without requiring us to muck with build deps?"

While in the neighborhood, I made some static fixBy* methods into instance methods on NullableAnnotationToUse. I can't remember if I've asked if you'd prefer this kind of change to be separated out. If you would, let me know, whether for this particular CL or only for the future.

PiperOrigin-RevId: 403260012