Don't add `@Nullable` to fields annotated with `@MonotonicNonNull`.
I accomplish this by treating `@MonotonicNonNull` as if it's identical to `@Nullable`.

That's not exactly true: Notably, [you can't assign `null` to a `@MonotonicNonNull` field](https://checkerframework.org/api/org/checkerframework/checker/nullness/qual/MonotonicNonNull.html). Yet this CL would now permit `foo = null` if `foo` is `@MonotonicNonNull`.

(To be fair, before this CL, we'd rewrite `foo` to have both the existing `@MonotonicNonNull` _and_ `@Nullable`. So it's not like we're breaking something that was truly "working" to begin with.)

In practice, I think this is a reasonable approach: `@MonotonicNonNull` is likely to be used only in CF-checked code, so it's fairly likely to be correct already.

Furthermore, `@MonotonicNonNull` is typically used on `private` fields. So we don't have to worry that other tools (like Kotlin) don't understand it. That eliminates one reason to prefer `@Nullable`.

PiperOrigin-RevId: 410926488