Encourage tightening types of variables & methods to convey more type information.
This subsumes MutableMethodReturnType, which I think make sense (they're broadly the same idea).

The scope is a bit broader. It would suggest replacing, say:

Iterable<Foo> foos() { return new ArrayList<>(); }

with:

List<Foo> foos() { return new ArrayList<>(); }

Because returning Iterable when you know it's a List is odd. Of course, if we think that's likely to prove controversial, I'm happy to back away.

PiperOrigin-RevId: 411763612