Check for an empty fix instead of a null fix.
(NullnessUtils always returns empty, never null.)

This change doesn't really matter yet: Currently, it should come up only with Nullness:OnlyIfAnnotationAlreadyInScope=true, which we set only during depot-wide cleanups, which we aren't planning for these particular refactorings yet. But it will matter once VoidMissingNullable uses a NullnessUtils.fix* method that returns an empty fix if the available annotation is not a type-use annotation when we need one.

This change brings these refactorings in line with EqualsMissingNullable and ReturnMissingNullable.

PiperOrigin-RevId: 413228181