Add a test that demonstrates that definitelyNullVars is not a good candidate for VisitorState.memoize(Supplier).
I had been hoping that an instance of VisitorState was created per compilation unit, but it's per compilation instance. And we don't want to use the known-null variables for one compilation unit when analyzing another. (OK, maybe that would be nice in very rare cases. But we definitely don't want to use the other compilation unit's known-null variables _in place of_ using our own :))

The background here is that I was wondering if hasDefinitelyNullBranch could compute definitelyNullVars on the fly instead of asking all its callers to precompute it. Presumably it still _could_, maybe even by using VisitorState.memoize(Supplier), but the Supplier would have to return a LoadingCache<CompilationUnitTree, ImmutableSet<VarSymbol>> or similar.

PiperOrigin-RevId: 413459090