Get type mappings in a more principled way
Reflectively accessing enclosing instances of inner classes using the
private synthetic `this$0` field is going to stop working in JDK 18:
https://bugs.openjdk.java.net/browse/JDK-8271623

We are currently relying on that as part of a hack to extra the
instantiation of type variables in method types. This change replaces
that logic by analyzing instantiated method types to match up types
with the corresponding type variables.

i.e., given a method `<A, B> List<B> foo(A a, B b)` and an invocation
with the type `(String,int)List<String>`, we would extract the mapping
`{A: String, B: int}`.

This replacement has the same behaviour for all of the existing tests
except for one case in `FutureReturnValueIgnored`. In that case the
reflective logic seems to be able to extract some intermediate work
javac does that can't be determined just by looking at the final
result of type inference. I think that's an acceptable regression,
because in the context of that check it's still rejecting the code,
just with one diagnostic instead of two.

PiperOrigin-RevId: 417828757