Short-circuit getEnclosingClass() if input is a ClassTree.
I believe this is the appropriate behavior with respect to the way identifier
resolution is currently implemented and used.  I suspect this has not been
noticed and fixed earlier because it is perhaps uncommon to attempt to
resolve identifier names against a path that starts with a ClassTree.

Currently, a Javadoc comment on a class cannot resolve the unqualified name of
a nested class using the ClassTree associated with the Javadoc comment.  e.g.:

/**
 * Bar. <-- Cannot resolve "Bar" using FindIdentifiers here.
 *
 * {@link Bar} <-- javac has no problem resolving "Bar" here.
 */
class Foo {
  class Bar {}
}

The Java compiler appears to resolve names inside of appropriate Javadoc tags
as though they were an expression scoped by the class the Javadoc comment is
bound to. This change brings that behavior to FindIdentifiers.

I need this as part of b/202156148 in order to find type names that should be
in {@link} tags.

PiperOrigin-RevId: 421356420