UnusedMethod: handle unused private non-zero-arg constructors.
Generating the fix is a bit of an ordeal. I've gone to some pains to a) leave behind a private constructor if we removed all the constructors (so the class doesn't become externally instantiable), and b) not suggest a fix if we're removing the last constructor that initialises final fields.

I think that might suffice, but I bet I missed something.

PiperOrigin-RevId: 422365990