Automated rollback of commit 9941ae5eff1d6e98de933bfe3210771e621a9a34.
*** Reason for rollback ***

Comment on unknown commit

*** Original change description ***

TypeCompatibilityUtils: regard a type as incompatible with _itself_ if it ends with "Builder" and doesn't override equals.

It's tempting to just say "types that don't override equals are not self-compatible", but that obviously doesn't work because of, for example, Enums. The heuristic of "ends with Builder" is to narrow down to types that someone might reasonably think have value semantics.

Flume: unknown commit

***

PiperOrigin-RevId: 424349932