For methods whose entire body is `return null`, begin adding `@Nullable` _if_ the method can't be overridden.
Anyone who calls such a method can definitely see a null result. In contrast, if the method were overridable, it might be only a "stub" that is overridden by all implementations to return non-null values.

(Also, update the comment about such "stub" methods to reflect that they are indeed common enough to be a problem, as discussed in https://github.com/google/error-prone/issues/2910#issuecomment-1022531264)

PiperOrigin-RevId: 427291930