Add an extra option `if_not` to `OnlyIfAnnotationAlreadyInScope`.
The options are now `true` (only add `@Nullable` if it is imported or otherwise in scope), `false` (always add `@Nullable` even if there is none in scope), and `if_not` (only add `@Nullable` if it is *not* imported or otherwise in scope).

Also make it so that enum-valued options can be specified case-insensitively.

The names might be improved.

PiperOrigin-RevId: 427551604