Restructure `UnusedReturnValueMatcher` a bit to differentiate between return values being used and unused return values being allowed for another reason.
As part of that, handle the case where a method has an unused return value of type `Void` as a reason to allow that rather than treating `void` and `Void` the same in some cases. Also fix an issue where a method reference like `future::get` where `future` is a `Future<Void>` wasn't treated as a method returning `Void`.

PiperOrigin-RevId: 427992486