Automated rollback of commit c971906bce43c771c172c43a9201d89d72fa6218.
*** Reason for rollback ***

b/220192531

*** Original change description ***

Introduce `ASTHelpers#isConsideredPrivate` to check whether a `Symbol` is private _or_ any of its owning elements are.

Also `ASTHelpers#isExternallyAccessible` to determine whether a `MethodSymbol` is both effectively private and not an override (and hence, ignoring reflection, unreachable outside the current file).

Doing this required fixing a bunch of (ahem) bugs, like not counting `private static Parcelable<...> CREATOR` as an unused variable.

***

PiperOrigin-RevId: 429418835