Treat `Thread.run` as `@DoNoCall`.
Callers often want `Thread.start`.

Yes, some do mean to run the work directly. But such code should probably already have a comment, so a suppression doesn't seem like a significant additional burden. And I assume that we're seeing a lot of survivor bias when we search for existing occurrences.

PiperOrigin-RevId: 461991043