Fix issue with CanIgnoreReturnValueSuggester.
When the method that produced the (ignorable) value to return:
  * Is a generic method
  * Whose upper bound was a supertype of the method we're analyzing

We would incorrectly choose to believe that the value coming from the
other method was substitutable for `this`.

We fix this by ensuring that we resolve the generic return type, and by
ensuring that the surrounding type *could* be returned by the method
we're analyzing.

PiperOrigin-RevId: 463119481