Add test to SuggesterTest.java that exposes a bug.
While creating unknown commit, I was suggested two similar auto-generated InlineMe annotations. One of them was fine, but the other had a bug. This cl here now adds two test cases to SuggesterTest which exposes the bug:

"US_ASCII" is a static import, so the generated InlineMe annotation should have a "staticImports" parameter. But in one of the two cases, it gets a "imports" parameter, which is wrong. Also, it is not fully qualified.

PiperOrigin-RevId: 468384719