Use a different (not incorrect this time) caching strategy for determining whether a `Tree` is part of an `import` or `package` statement.
The previous caching strategy was incorrect because `VisitorState.memoize` memoizes across all source files in a compilation, not over each source file, which meant that the the import/package ranges from the _first_ source file that was being processed when they were requested were used for _all_ the source files.

The new code caches a map from source file to import-and-package ranges.

Tested:
    Sample TGP: []
PiperOrigin-RevId: 468767587