Avoid name clashes when MustBeClosedChecker introduces multiple variables
This is sorta a revert of https://github.com/google/error-prone/commit/b19ba52e167778ba509db9f9953cf13600db8e5f. It turns out I was wrong about not needing this feature. But having removed it at least gave me a better understanding of what's going on in this file, so I could implement a non-buggy version of it (demonstrated by the new test).

PiperOrigin-RevId: 474682278