Adds tests and refactors existing tests for clarity.
As currently implemented, this bugpattern checker only handles an extremely narrow issue set and likely gives developers much more confidence in Optional safety than is actually provided. If we paste the same two files into the Checker Framework's live demo for the Optional checker, it passes on all test cases except that it still has an (understandably) false-positive for the OptionalNotPresentNegativeCases.getWhenTestedSafe_equals.

I would recommend we highlight the shortcomings of this checker very clearly.

PiperOrigin-RevId: 477896538