refactor: refactor bad smell UnusedLabel
# Repairing Code Style Issues
<!-- laughing-train-refactor -->
## UnusedLabel
A label is a named statement that can be used to jump to it from other parts of the code. If a label is not used, it is better to remove it.

<!-- fingerprint:-1777244990 -->
# Repairing Code Style Issues
* UnusedLabel (1)

Fixes #3519

COPYBARA_INTEGRATE_REVIEW=https://github.com/google/error-prone/pull/3519 from MartinWitt:laughing-train/Qodana/-331951670 c4fdf9c21b990e6ea23950865785845bff03e370
PiperOrigin-RevId: 499314932