Adjust `PlaceholderUnificationVisitor` so it works with older and newer `TreeMaker` APIs.
The type of the first parameter of `TreeMaker.ForeachLoop` changed from `JCVariableDecl` to `JCTree` to reflect the evolution of the language. We use `MethodHandle` so we can work with either version.

This change works with the default config; with `--config=java_runtime_head`; and with `--config=java_language_head` if certain other unrelated problems are patched.

PiperOrigin-RevId: 507920350