Adjust the regex for detecting empty varargs parameters.
In the test, we end up checking the string `Client.execute2(format, args)` against the pattern `\b(?:,\s*)?\Qargs\E\b([^(])`. That's kind of ambiguous: the first `\b` (word boundary) could match the end of `format`, which is what's wanted, or it could match the start of `args`, with `(?:,\s*)?` matching nothing. Apparently the latter is what happens, but if we remove the first `\b` we get the match we want.

It would probably be better to rewrite this in a less obscure way.

#inlineme

PiperOrigin-RevId: 508131197