MemberName: check the validity of acronym camelcasing.
As a heuristic, assume that three upperacase letters in a row are probably an acronym, or two if they're at the end of a member name.

e.g. `getRPCPolicy` is suspicious. I think `rPCPolicy` is unlikely to be an issue, because it looks so bad I can't imagine people writing it.

Partial flume over ErrorProne: unknown commit. Flume everywhere (huge): unknown commit[1].

[1] I added the serialVersionUID exemption after that flume.

PiperOrigin-RevId: 536708104