When an inlined method refers to method arguments in the true or false branch of a conditional statement, refuse to allow it to be inlined.
Before calling the method in question, the caller has definitely
executed the expressions that yield the values passed to the method.

However, after inlining the body of the method into the callsite, one
of the two expressions will no longer be executed, leading to a subtle
behavior change if the un-executed expression was side-effecting.

PiperOrigin-RevId: 617624766