Automated rollback of commit a3fdf64120fd254d520e6fb98eaf0cb8d1e9b17e.
*** Reason for rollback ***

Requires a large number of fixes throughout the depot. I am not sure that that was intended, but if it was then those fixes should be made separately from the regularly-scheduled JavaBuilder release.

A *subset* of the fixes, before I gave up, is here: unknown commit.

*** Original change description ***

Generalize InjectScopeOrQualifierAnnotationRetention to MissingRuntimeRetention

***

PiperOrigin-RevId: 646486667