Test that `FieldMissingNullableTest` looks at record components when searching for or adding annotations.
(It does.)

There is still more that we could do for records, like looking for tests of `someComponent() == null` that use the _method_ `someComponent()` instead of the field. But this seems fine for now, especially since code within the record class is presumably likely to use the field where possible.

PiperOrigin-RevId: 647255480