Add support in ErrorProne for a new `ThreadSafeTypeParamerter` annotation that is replacing `ThreadSafe.TypeParameter`.
A follow-up CL will actually create the annotation once ErrorProne has this CL released to avoid premature dependencies on the new annotation.

This is the first step in open sourcing the threadsafe equivalent of `ImmutableTypeParameter`. After this an LSC will be performed to replace all usages of `ThreadSafe.TypeParameter` with `ThreadSafeTypeParameter` after which and remaining references to the former (e.g. in messages) will be removed.

PiperOrigin-RevId: 651816128