Add Error Prone check for unnecessary boxed types in AutoValue classes.
This check will warn users when they are using boxed types in their AutoValue classes that are not Nullable. This is because boxed types are not necessary in AutoValue classes, and they can actually be harmful because they can cause unnecessary boxing and unboxing.

Clean up reference: unknown commit

This check is currently enabled as a warning, but it can be made into an error in the future.

PiperOrigin-RevId: 658493053