Prepare for the removal of `TypeTag.UNKNOWN`
The constant was removed as part of https://bugs.openjdk.org/browse/JDK-8339296

This is a minimal fix to avoid breaking for upcoming javac versions. I suspect this logic isn't actually necessary, we aren't supposed to process ASTs with errors and there have been a variety of improvements in how that's handled since this was added.

PiperOrigin-RevId: 690616927