add values() to EnumOrdinal check
This check is in JavaCodeClarity, only runs on changed lines.

While testing, discovered that it would flag a lot of instances where `values()[` was being invoked from within the same class. I wasn't sure if that was a reasonable thing to do, seems like it might be a valid internal design choice. See the 'negative_enumValues_internalCall' test case as an example. Or see some of the instances in unknown commit

So I added an additional check of the enclosing class to ignore internal usage. That flagged fewer things: unknown commit

PiperOrigin-RevId: 705970604