Restore a curated list of GuardedBy annotations to heed.
To paraphrase offline discussion: in an ideal world, we'd like all our checks to be consistent and only obey just the EP annotations. That's currently true for Immutable and ThreadSafe, but not for GuardedBy. It'd be nice to either only obey EP's for everything, or obey everything based on simple name. A compromise is to curate a list.

PiperOrigin-RevId: 732862069