[PatternMatchingInstanceof] Make negated matches optional
This introduces a configuration option that allows one to disable
the replacement of casts that are implied by a negated instanceof
expression and are not within the same expression as the instanceof.
This affects casts in else branches and those below an if statement
where the then branch always returns from the method.
Casts where the readability benefit of a pattern-maching instanceof
is clear, i.e. those within the then branch of an if
and those inside the same expression, are still replaced.

The default of the option is true, which keeps the current behavior.

Fixes #4925