Yet more PatternMatchingInstanceof findings.
Generated with unknown commit, but regarding all method invocations as pure. (These findings were good enough that I almost think that's a good heuristic for this check! If you're doing an instance of and then a cast, you almost certainly are at least pretending that method is pure.)

I made follow-on simplifications.

PiperOrigin-RevId: 742734172