Tolerate default cases in switches as being present to handle version skew
If the default branch of a switch is commented with the phrase "skew", we'll now
assume that the developer added it explicitly to handle version skew concerns.

As a result:

 * `UnnecessaryDefaultInEnumSwitch` will no longer complain about the `default`
   branch being unnecessary if it's already otherwise exhaustive.
 * `MissingCasesInEnumSwitch` will continue to check to ensure all cases are
   handled despite the default branch being present.

This is intended to provide better safety for cases where developers have
legitimate concerns about version skew, but don't want to give up the benefit
of checking to ensure all cases are covered.

If a user doesn't want to be exhaustive then they should not comment the default
branch as being present for skew as it's clearly intended to handle more cases
than that.

PiperOrigin-RevId: 748697198