Warn on qualifiers used in places that are not injection points.
This is a check I've wanted to write for a long time. IntelliJ used to copy qualifiers over when extracting methods, which is very annoying.

As you might expect, given how easy it is for this to happen, the Flume results are enormous: http://unknown commit

`UnnecessaryQualifier-findings 87,665`

There are some interesting sources of false positives, and annoyingly the entire first page of the Flume results are. I've mailed fixes for them.

  1. Incorrect use of Qualifier on annotations. Some people seem to think all annotations are qualifiers!
  2. Curious custom DI frameworks. I've exempted some in the list of classes. We could add more if needed; we could also consider a meta-annotation. (It would be hard to get the semantics for an annotation right, though.)

PiperOrigin-RevId: 755408778