Extend MissingCasesInEnumSwitch to switch expressions if the default is only for skew
Normally Javac will ensure that switch expressions are exhaustive, so Error
Prone can typically just skip them. However, if someone has a default branch
that they only intend to have present for skew, it may still be useful to ensure
that all the enum values are covered.

Therefore, if a switch expression has a default branch and it's marked as being
for skew, we'll now check that it's still otherwise exhaustive.

PiperOrigin-RevId: 758713021